CHANGELOG entry + KDocs updates
